Strings and Text
What is a String?
A string is usually a bit of text you want to display to someone or "export" out of the program you are writing. Python know that you want something to be a string when you put them in double quotes (" ")  or single quotes ( ' ' ).

Python has another type of formatting using the .format( ) syntax.